+++ includes/cipher.h
@@ -34,6 +34,6 @@ public:
 
 	virtual void test(){};
 
-	virtual uint8 apply_s(uint8 input, int box) = 0;
+	virtual uint8 apply_s(uint16 input, int box) = 0;
 
 };
